Break temporary retain cycle in system presentation style (#40)
By retaining the entire `context` in the captured observer closure, we're implicitly capturing the `context.presenting` `UIViewController` for the duration of the transition until the `DismissalTransitionDidEndNotification` notification is fired.

If the `DismissalTransitionDidEndNotification` is never triggered, e.g. in the case of early deallocation following presentation, this could lead to the presenting view controller being retained indefinitely.